Create retry-strategy package and split different implementation in separate files
