+++ src/non-resumable-subscription.ts
@@ -1,6 +1,6 @@
 import { Logger } from './logger';
 import { TokenProvider } from './token-provider';
-import { RetryStrategy } from './retry-strategy';
+import { RetryStrategy } from './retry-strategy/retry-strategy';
 import { ErrorResponse, Headers } from './base-client';
 import { BaseSubscription, BaseSubscriptionConstruction, SubscriptionEvent } from './base-subscription';
